introduced error and repair functions
centralized the errors into new class, TessellationError.
changed a bunch of little things, but more repair functions are needed.
The TessellationError.Report() method gives us some ideas of how to
proceed.

+++ TessellationAndVoxelizationGeometryLibrary/Boolean Operations/Slice3.cs
@@ -226,7 +226,7 @@ namespace TVGL.Boolean_Operations
             /// <summary>
             /// Original edge checksum reference
             /// </summary>
-            public int EdgeReference;
+            public long EdgeReference;
 
             internal StraddleEdge(Edge edge, Flat plane)
             {
